Now works through most of the platforms! (browserify, NodeJS, webpack, commonJS, AMD...)
