+++ package.json
@@ -2,16 +2,19 @@
   "name": "game-time-accumulator",
   "version": "1.0.0",
   "description": "game-time-accumulator class for creating a game",
-  "main": "dist/GameTimeAccumulator.js",
+  "main": "dist/game-time-accumulator.js",
   "dependencies": {},
   "devDependencies": {
-    "babelify": "^6.1.3",
-    "browserify": "^11.0.1",
-    "raf": "^3.1.0"
+    "babel-core": "^5.8.22",
+    "babel-loader": "^5.3.2",
+    "path": "^0.11.14",
+    "raf": "^3.1.0",
+    "uglify-js": "^2.4.24",
+    "webpack": "^1.11.0"
   },
   "engines": {},
   "scripts": {
-    "build": "browserify src/GameTimeAccumulator.js -t babelify --outfile dist/GameTimeAccumulator.js"
+    "build": "webpack && uglifyjs dist/game-time-accumulator.js -o dist/game-time-accumulator.min.js"
   },
   "repository": {
     "type": "git",
@@ -24,7 +27,7 @@
     "es6"
   ],
   "author": "Yuval Saraf",
-  "license": "ISC",
+  "license": "MIT",
   "bugs": {
     "url": "https://github.com/yuvalsaraf/game-time-accumulator/issues"
   },
